# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type StudyModality = typeof StudyModality[keyof typeof StudyModality];


export const StudyModality = {
  xray: 'xray',
  mri: 'mri',
} as const;

export type StudyStatus = typeof StudyStatus[keyof typeof StudyStatus];


export const StudyStatus = {
  unread: 'unread',
  review: 'review',
  finalized: 'finalized',
} as const;

export type StudyPriority = typeof StudyPriority[keyof typeof StudyPriority];


export const StudyPriority = {
  routine: 'routine',
  urgent: 'urgent',
  stat: 'stat',
} as const;

export type FindingSeverity = typeof FindingSeverity[keyof typeof FindingSeverity];


export const FindingSeverity = {
  normal: 'normal',
  low: 'low',
  moderate: 'moderate',
  high: 'high',
} as const;

export interface Finding {
  id: string;
  label: string;
  confidence: number;
  severity: FindingSeverity;
  description: string;
  heatmapRegion: string;
}

export type ReportStatus = typeof ReportStatus[keyof typeof ReportStatus];


export const ReportStatus = {
  draft: 'draft',
  finalized: 'finalized',
} as const;

export interface Report {
  indication: string;
  findings: string;
  impression: string;
  status: ReportStatus;
  updatedAt?: string;
}

export interface Study {
  id: string;
  patientId: string;
  patientName: string;
  accessionNumber: string;
  modality: StudyModality;
  bodyPart: string;
  studyDate: string;
  status: StudyStatus;
  priority: StudyPriority;
  imageCount: number;
  imageUrl: string;
  thumbnailUrl?: string;
  /** @nullable */
  modelVersion?: string | null;
  /** @nullable */
  inferenceTimeMs?: number | null;
  findings: Finding[];
  report: Report;
}

export type StudyInputModality = typeof StudyInputModality[keyof typeof StudyInputModality];


export const StudyInputModality = {
  xray: 'xray',
  mri: 'mri',
} as const;

export interface StudyInput {
  patientId: string;
  patientName: string;
  accessionNumber: string;
  modality: StudyInputModality;
  bodyPart: string;
  imageUrl: string;
  imageCount: number;
}

export type AnalysisResultStatus = typeof AnalysisResultStatus[keyof typeof AnalysisResultStatus];


export const AnalysisResultStatus = {
  complete: 'complete',
  processing: 'processing',
} as const;

export interface AnalysisResult {
  studyId: string;
  modelName: string;
  modelVersion: string;
  status: AnalysisResultStatus;
  findings: Finding[];
  completedAt: string;
}

export type DashboardSummaryModalityMixItem = {
  label: string;
  value: number;
};

export type DashboardSummaryActivityItem = {
  label: string;
  value: number;
};

export interface DashboardSummary {
  studiesToday: number;
  pendingReview: number;
  statStudies: number;
  finalizedToday: number;
  avgTurnaroundMinutes: number;
  modalityMix: DashboardSummaryModalityMixItem[];
  activity: DashboardSummaryActivityItem[];
}

export interface AuditEvent {
  id: string;
  action: string;
  actor: string;
  timestamp: string;
  detail: string;
}

export interface ReportUpdate {
  indication: string;
  findings: string;
  impression: string;
  finalize: boolean;
}

export type ListStudiesParams = {
status?: ListStudiesStatus;
modality?: ListStudiesModality;
search?: string;
};

export type ListStudiesStatus = typeof ListStudiesStatus[keyof typeof ListStudiesStatus];


export const ListStudiesStatus = {
  all: 'all',
  unread: 'unread',
  review: 'review',
  finalized: 'finalized',
} as const;

export type ListStudiesModality = typeof ListStudiesModality[keyof typeof ListStudiesModality];


export const ListStudiesModality = {
  all: 'all',
  xray: 'xray',
  mri: 'mri',
} as const;