# lib/api-zod/src/generated/types/analysisResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AnalysisResultStatus } from './analysisResultStatus';
import type { Finding } from './finding';

export interface AnalysisResult {
  studyId: string;
  modelName: string;
  modelVersion: string;
  status: AnalysisResultStatus;
  findings: Finding[];
  completedAt: string;
}