# lib/api-zod/src/generated/types/auditEvent.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AuditEvent {
  id: string;
  action: string;
  actor: string;
  timestamp: string;
  detail: string;
}