# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardSummaryActivityItem } from './dashboardSummaryActivityItem';
import type { DashboardSummaryModalityMixItem } from './dashboardSummaryModalityMixItem';

export interface DashboardSummary {
  studiesToday: number;
  pendingReview: number;
  statStudies: number;
  finalizedToday: number;
  avgTurnaroundMinutes: number;
  modalityMix: DashboardSummaryModalityMixItem[];
  activity: DashboardSummaryActivityItem[];
}