# lib/api-zod/src/generated/types/finding.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FindingSeverity } from './findingSeverity';

export interface Finding {
  id: string;
  label: string;
  confidence: number;
  severity: FindingSeverity;
  description: string;
  heatmapRegion: string;
}