# lib/api-zod/src/generated/types/report.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ReportStatus } from './reportStatus';

export interface Report {
  indication: string;
  findings: string;
  impression: string;
  status: ReportStatus;
  updatedAt?: string;
}