# lib/api-zod/src/generated/types/reportUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ReportUpdate {
  indication: string;
  findings: string;
  impression: string;
  finalize: boolean;
}