# lib/api-zod/src/generated/types/study.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Finding } from './finding';
import type { Report } from './report';
import type { StudyModality } from './studyModality';
import type { StudyPriority } from './studyPriority';
import type { StudyStatus } from './studyStatus';

export interface Study {
  id: string;
  patientId: string;
  patientName: string;
  accessionNumber: string;
  modality: StudyModality;
  bodyPart: string;
  studyDate: string;
  status: StudyStatus;
  priority: StudyPriority;
  imageCount: number;
  imageUrl: string;
  thumbnailUrl?: string;
  /** @nullable */
  modelVersion?: string | null;
  /** @nullable */
  inferenceTimeMs?: number | null;
  findings: Finding[];
  report: Report;
}